+++ src/KubernetesClient/Enums/WatchEventType.cs
@@ -1,0 +1,32 @@
+﻿using System.Runtime.Serialization;
+
+namespace k8s.Enums
+{
+    /// <summary>Describes the type of a watch event.</summary>
+    public enum WatchEventType
+    {
+        /// <summary>Emitted when an object is created, modified to match a watch's filter, or when a watch is first opened.</summary>
+        [EnumMember(Value = "ADDED")]
+        Added,
+
+        /// <summary>Emitted when an object is modified.</summary>
+        [EnumMember(Value = "MODIFIED")]
+        Modified,
+
+        /// <summary>Emitted when an object is deleted or modified to no longer match a watch's filter.</summary>
+        [EnumMember(Value = "DELETED")]
+        Deleted,
+
+        /// <summary>Emitted when an error occurs while watching resources. Most commonly, the error is 410 Gone which indicates that
+        /// the watch resource version was outdated and events were probably lost. In that case, the watch should be restarted.
+        /// </summary>
+        [EnumMember(Value = "ERROR")]
+        Error,
+
+        /// <summary>Bookmarks may be emitted periodically to update the resource version. The object will
+        /// contain only the resource version.
+        /// </summary>
+        [EnumMember(Value = "BOOKMARK")]
+        Bookmark,
+    }
+}
