+++ src/KubernetesClient/IKubernetes.Watch.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using k8s.Enums;
 
 namespace k8s
 {
