+++ src/KubernetesClient/Kubernetes.Watch.cs
@@ -7,6 +7,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using k8s.Enums;
 
 namespace k8s
 {
